+++ node/src/Transport.ts
@@ -402,7 +402,7 @@ export abstract class Transport<
 		this.internal = internal;
 		this.#data = data;
 		this.channel = channel;
-		this.#appData = appData || ({} as TransportAppData);
+		this.#appData = appData ?? ({} as TransportAppData);
 		this.#getRouterRtpCapabilities = getRouterRtpCapabilities;
 		this.getProducerById = getProducerById;
 		this.getDataProducerById = getDataProducerById;
@@ -635,6 +635,7 @@ export abstract class Transport<
 	 *
 	 * @abstract
 	 */
+	// eslint-disable-next-line @typescript-eslint/require-await
 	async dump(): Promise<any> {
 		// Should not happen.
 		throw new Error('method implemented in the subclass');
@@ -645,6 +646,7 @@ export abstract class Transport<
 	 *
 	 * @abstract
 	 */
+	// eslint-disable-next-line @typescript-eslint/require-await
 	async getStats(): Promise<any[]> {
 		// Should not happen.
 		throw new Error('method implemented in the subclass');
@@ -655,7 +657,7 @@ export abstract class Transport<
 	 *
 	 * @abstract
 	 */
-	// eslint-disable-next-line @typescript-eslint/no-unused-vars
+	// eslint-disable-next-line @typescript-eslint/no-unused-vars, @typescript-eslint/require-await
 	async connect(params: any): Promise<void> {
 		// Should not happen.
 		throw new Error('method implemented in the subclass');
@@ -761,11 +763,7 @@ export abstract class Transport<
 		if (this.constructor.name !== 'PipeTransport') {
 			// If CNAME is given and we don't have yet a CNAME for Producers in this
 			// Transport, take it.
-			if (
-				!this.#cnameForProducers &&
-				clonedRtpParameters.rtcp &&
-				clonedRtpParameters.rtcp.cname
-			) {
+			if (!this.#cnameForProducers && clonedRtpParameters.rtcp?.cname) {
 				this.#cnameForProducers = clonedRtpParameters.rtcp.cname;
 			}
 			// Otherwise if we don't have yet a CNAME for Producers and the RTP
@@ -795,7 +793,7 @@ export abstract class Transport<
 			rtpMapping
 		);
 
-		const producerId = id || utils.generateUUIDv4();
+		const producerId = id ?? utils.generateUUIDv4();
 		const requestOffset = createProduceRequest({
 			builder: this.channel.bufferBuilder,
 			producerId,
@@ -827,7 +825,7 @@ export abstract class Transport<
 			consumableRtpParameters,
 		};
 
-		const producer = new Producer<ProducerAppData>({
+		const producer: Producer<ProducerAppData> = new Producer({
 			internal: {
 				...this.internal,
 				producerId,
@@ -954,7 +952,7 @@ export abstract class Transport<
 			type: pipe ? 'pipe' : (producer.type as ConsumerType),
 		};
 
-		const consumer = new Consumer<ConsumerAppData>({
+		const consumer: Consumer<ConsumerAppData> = new Consumer({
 			internal: {
 				...this.internal,
 				consumerId,
@@ -968,10 +966,7 @@ export abstract class Transport<
 			preferredLayers: status.preferredLayers
 				? {
 						spatialLayer: status.preferredLayers.spatialLayer,
-						temporalLayer:
-							status.preferredLayers.temporalLayer !== null
-								? status.preferredLayers.temporalLayer
-								: undefined,
+						temporalLayer: status.preferredLayers.temporalLayer ?? undefined,
 					}
 				: undefined,
 		});
@@ -1034,7 +1029,7 @@ export abstract class Transport<
 			}
 		}
 
-		const dataProducerId = id || utils.generateUUIDv4();
+		const dataProducerId = id ?? utils.generateUUIDv4();
 		const requestOffset = createProduceDataRequest({
 			builder: this.channel.bufferBuilder,
 			dataProducerId,
@@ -1059,7 +1054,7 @@ export abstract class Transport<
 
 		const dump = parseDataProducerDumpResponse(produceDataResponse);
 
-		const dataProducer = new DataProducer<DataProducerAppData>({
+		const dataProducer: DataProducer<DataProducerAppData> = new DataProducer({
 			internal: {
 				...this.internal,
 				dataProducerId,
@@ -1194,7 +1189,7 @@ export abstract class Transport<
 
 		const dump = parseDataConsumerDumpResponse(consumeDataResponse);
 
-		const dataConsumer = new DataConsumer<DataConsumerAppData>({
+		const dataConsumer: DataConsumer<DataConsumerAppData> = new DataConsumer({
 			internal: {
 				...this.internal,
 				dataConsumerId,
@@ -1338,10 +1333,6 @@ export function parseSctpState(fbsSctpState: FbsSctpState): SctpState {
 
 		case FbsSctpState.CLOSED: {
 			return 'closed';
-		}
-
-		default: {
-			throw new TypeError(`invalid SctpState: ${fbsSctpState}`);
 		}
 	}
 }
@@ -1530,7 +1521,7 @@ export function parseTransportTraceEventData(
 				timestamp: Number(trace.timestamp()),
 				direction:
 					trace.direction() === FbsTraceDirection.DIRECTION_IN ? 'in' : 'out',
-				info: parseBweTraceInfo(info!),
+				info: parseBweTraceInfo(info),
 			};
 		}
 
@@ -1572,10 +1563,6 @@ function transportTraceEventTypeToFbs(
 
 		case 'bwe': {
 			return FbsTransport.TraceEventType.BWE;
-		}
-
-		default: {
-			throw new TypeError(`invalid TransportTraceEventType: ${eventType}`);
 		}
 	}
 }