+++ node/src/enhancedEvents.ts
@@ -113,10 +113,12 @@ export class EnhancedEventEmitter<
 		return super.listenerCount(eventName);
 	}
 
+	// eslint-disable-next-line @typescript-eslint/no-unsafe-function-type
 	listeners<K extends keyof E & string>(eventName: K): Function[] {
 		return super.listeners(eventName);
 	}
 
+	// eslint-disable-next-line @typescript-eslint/no-unsafe-function-type
 	rawListeners<K extends keyof E & string>(eventName: K): Function[] {
 		return super.rawListeners(eventName);
 	}