+++ node/src/test/test-Consumer.ts
@@ -195,13 +195,13 @@ const ctx: TestContext = {
 				},
 				{
 					kind: 'audio',
-					uri: 'http://www.webrtc.org/experiments/rtp-hdrext/abs-send-time', // eslint-disable-line max-len
+					uri: 'http://www.webrtc.org/experiments/rtp-hdrext/abs-send-time',
 					preferredId: 4,
 					preferredEncrypt: false,
 				},
 				{
 					kind: 'video',
-					uri: 'http://www.webrtc.org/experiments/rtp-hdrext/abs-send-time', // eslint-disable-line max-len
+					uri: 'http://www.webrtc.org/experiments/rtp-hdrext/abs-send-time',
 					preferredId: 4,
 					preferredEncrypt: false,
 				},
@@ -454,7 +454,7 @@ test('transport.consume() succeeds', async () => {
 	});
 	expect(videoPipeConsumer.preferredLayers).toBeUndefined();
 	expect(videoPipeConsumer.currentLayers).toBeUndefined();
-	expect(videoPipeConsumer.appData).toBeUndefined;
+	expect(videoPipeConsumer.appData).toEqual({});
 
 	const dump2 = await ctx.router!.dump();
 
@@ -818,11 +818,11 @@ test('consumer.pause() and resume() succeed', async () => {
 
 	// Even if we don't await for pause()/resume() completion, the observer must
 	// fire 'pause' and 'resume' events if state was the opposite.
-	audioConsumer.pause();
-	audioConsumer.resume();
-	audioConsumer.pause();
-	audioConsumer.pause();
-	audioConsumer.pause();
+	void audioConsumer.pause();
+	void audioConsumer.resume();
+	void audioConsumer.pause();
+	void audioConsumer.pause();
+	void audioConsumer.pause();
 	await audioConsumer.resume();
 
 	expect(onObserverPause).toHaveBeenCalledTimes(3);
@@ -909,22 +909,22 @@ test('consumer.setPreferredLayers() with wrong arguments rejects with TypeError'
 		rtpCapabilities: ctx.consumerDeviceCapabilities,
 	});
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await expect(videoConsumer.setPreferredLayers({})).rejects.toThrow(TypeError);
 
 	await expect(
-		// @ts-ignore
+		// @ts-expect-error --- Testing purposes.
 		videoConsumer.setPreferredLayers({ foo: '123' })
 	).rejects.toThrow(TypeError);
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await expect(videoConsumer.setPreferredLayers('foo')).rejects.toThrow(
 		TypeError
 	);
 
 	// Missing spatialLayer.
 	await expect(
-		// @ts-ignore
+		// @ts-expect-error --- Testing purposes.
 		videoConsumer.setPreferredLayers({ temporalLayer: 2 })
 	).rejects.toThrow(TypeError);
 }, 2000);
@@ -945,12 +945,12 @@ test('consumer.setPriority() with wrong arguments rejects with TypeError', async
 		rtpCapabilities: ctx.consumerDeviceCapabilities,
 	});
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await expect(videoConsumer.setPriority()).rejects.toThrow(TypeError);
 
 	await expect(videoConsumer.setPriority(0)).rejects.toThrow(TypeError);
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await expect(videoConsumer.setPriority('foo')).rejects.toThrow(TypeError);
 }, 2000);
 
@@ -981,7 +981,7 @@ test('consumer.enableTraceEvent() succeed', async () => {
 
 	expect(dump2.traceEventTypes).toEqual(expect.arrayContaining([]));
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await audioConsumer.enableTraceEvent(['nack', 'FOO', 'fir']);
 
 	const dump3 = await audioConsumer.dump();
@@ -1003,16 +1003,16 @@ test('consumer.enableTraceEvent() with wrong arguments rejects with TypeError', 
 		rtpCapabilities: ctx.consumerDeviceCapabilities,
 	});
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await expect(audioConsumer.enableTraceEvent(123)).rejects.toThrow(TypeError);
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await expect(audioConsumer.enableTraceEvent('rtp')).rejects.toThrow(
 		TypeError
 	);
 
 	await expect(
-		// @ts-ignore
+		// @ts-expect-error --- Testing purposes.
 		audioConsumer.enableTraceEvent(['fir', 123.123])
 	).rejects.toThrow(TypeError);
 }, 2000);
@@ -1144,7 +1144,7 @@ test('Consumer methods reject if closed', async () => {
 
 	await expect(audioConsumer.resume()).rejects.toThrow(Error);
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await expect(audioConsumer.setPreferredLayers({})).rejects.toThrow(Error);
 
 	await expect(audioConsumer.setPriority(2)).rejects.toThrow(Error);