+++ node/src/test/test-DataProducer.ts
@@ -146,7 +146,7 @@ test('webRtcTransport1.produceData() with wrong arguments rejects with TypeError
 	// Missing or empty sctpStreamParameters.streamId.
 	await expect(
 		ctx.webRtcTransport1!.produceData({
-			// @ts-ignore
+			// @ts-expect-error --- Testing purposes.
 			sctpStreamParameters: { foo: 'foo' },
 		})
 	).rejects.toThrow(TypeError);
@@ -271,11 +271,11 @@ test('dataProducer.pause() and resume() succeed', async () => {
 
 	// Even if we don't await for pause()/resume() completion, the observer must
 	// fire 'pause' and 'resume' events if state was the opposite.
-	dataProducer1.pause();
-	dataProducer1.resume();
-	dataProducer1.pause();
-	dataProducer1.pause();
-	dataProducer1.pause();
+	void dataProducer1.pause();
+	void dataProducer1.resume();
+	void dataProducer1.pause();
+	void dataProducer1.pause();
+	void dataProducer1.pause();
 	await dataProducer1.resume();
 
 	expect(onObserverPause).toHaveBeenCalledTimes(3);