+++ node/src/test/test-PlainTransport.ts
@@ -173,7 +173,7 @@ test('router.createPlainTransport() succeeds', async () => {
 }, 2000);
 
 test('router.createPlainTransport() with wrong arguments rejects with TypeError', async () => {
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await expect(ctx.router!.createPlainTransport({})).rejects.toThrow(TypeError);
 
 	await expect(
@@ -191,7 +191,7 @@ test('router.createPlainTransport() with wrong arguments rejects with TypeError'
 	).rejects.toThrow(TypeError);
 
 	await expect(
-		// @ts-ignore
+		// @ts-expect-error --- Testing purposes.
 		ctx.router!.createPlainTransport({ listenIp: ['127.0.0.1'] })
 	).rejects.toThrow(TypeError);
 
@@ -204,7 +204,7 @@ test('router.createPlainTransport() with wrong arguments rejects with TypeError'
 	await expect(
 		ctx.router!.createPlainTransport({
 			listenInfo: { protocol: 'udp', ip: '127.0.0.1' },
-			// @ts-ignore
+			// @ts-expect-error --- Testing purposes.
 			appData: 'NOT-AN-OBJECT',
 		})
 	).rejects.toThrow(TypeError);
@@ -237,7 +237,7 @@ test('router.createPlainTransport() with enableSrtp succeeds', async () => {
 		plainTransport.connect({
 			ip: '127.0.0.2',
 			port: 9999,
-			// @ts-ignore
+			// @ts-expect-error --- Testing purposes.
 			srtpParameters: 1,
 		})
 	).rejects.toThrow(TypeError);
@@ -247,7 +247,7 @@ test('router.createPlainTransport() with enableSrtp succeeds', async () => {
 		plainTransport.connect({
 			ip: '127.0.0.2',
 			port: 9999,
-			// @ts-ignore
+			// @ts-expect-error --- Testing purposes.
 			srtpParameters: {
 				keyBase64: 'ZnQ3eWJraDg0d3ZoYzM5cXN1Y2pnaHU5NWxrZTVv',
 			},
@@ -259,7 +259,7 @@ test('router.createPlainTransport() with enableSrtp succeeds', async () => {
 		plainTransport.connect({
 			ip: '127.0.0.2',
 			port: 9999,
-			// @ts-ignore
+			// @ts-expect-error --- Testing purposes.
 			srtpParameters: {
 				cryptoSuite: 'AES_CM_128_HMAC_SHA1_80',
 			},
@@ -272,7 +272,7 @@ test('router.createPlainTransport() with enableSrtp succeeds', async () => {
 			ip: '127.0.0.2',
 			port: 9999,
 			srtpParameters: {
-				// @ts-ignore
+				// @ts-expect-error --- Testing purposes.
 				cryptoSuite: 'FOO',
 				keyBase64: 'ZnQ3eWJraDg0d3ZoYzM5cXN1Y2pnaHU5NWxrZTVv',
 			},
@@ -285,7 +285,7 @@ test('router.createPlainTransport() with enableSrtp succeeds', async () => {
 			ip: '127.0.0.2',
 			port: 9999,
 			srtpParameters: {
-				// @ts-ignore
+				// @ts-expect-error --- Testing purposes.
 				cryptoSuite: 123,
 				keyBase64: 'ZnQ3eWJraDg0d3ZoYzM5cXN1Y2pnaHU5NWxrZTVv',
 			},
@@ -299,7 +299,7 @@ test('router.createPlainTransport() with enableSrtp succeeds', async () => {
 			port: 9999,
 			srtpParameters: {
 				cryptoSuite: 'AES_CM_128_HMAC_SHA1_80',
-				// @ts-ignore
+				// @ts-expect-error --- Testing purposes.
 				keyBase64: [],
 			},
 		})
@@ -482,7 +482,7 @@ test('plainTransport.connect() with wrong arguments rejects with TypeError', asy
 		plainTransport.connect({
 			ip: '127.0.0.1',
 			port: 1234,
-			// @ts-ignore
+			// @ts-expect-error --- Testing purposes.
 			__rtcpPort: 1235,
 		})
 	).rejects.toThrow(TypeError);
@@ -490,7 +490,7 @@ test('plainTransport.connect() with wrong arguments rejects with TypeError', asy
 	await expect(
 		plainTransport.connect({
 			ip: '127.0.0.1',
-			// @ts-ignore
+			// @ts-expect-error --- Testing purposes.
 			__port: 'chicken',
 			rtcpPort: 1235,
 		})