+++ node/src/test/test-Producer.ts
@@ -271,9 +271,9 @@ test('webRtcTransport1.produce() without header extensions and rtcp succeeds', a
 test('webRtcTransport1.produce() with wrong arguments rejects with TypeError', async () => {
 	await expect(
 		ctx.webRtcTransport1!.produce({
-			// @ts-ignore
+			// @ts-expect-error --- Testing purposes.
 			kind: 'chicken',
-			// @ts-ignore
+			// @ts-expect-error --- Testing purposes.
 			rtpParameters: {},
 		})
 	).rejects.toThrow(TypeError);
@@ -281,7 +281,7 @@ test('webRtcTransport1.produce() with wrong arguments rejects with TypeError', a
 	await expect(
 		ctx.webRtcTransport1!.produce({
 			kind: 'audio',
-			// @ts-ignore
+			// @ts-expect-error --- Testing purposes.
 			rtpParameters: {},
 		})
 	).rejects.toThrow(TypeError);
@@ -293,7 +293,7 @@ test('webRtcTransport1.produce() with wrong arguments rejects with TypeError', a
 			rtpParameters: {
 				codecs: [],
 				headerExtensions: [],
-				// @ts-ignore
+				// @ts-expect-error --- Testing purposes.
 				encodings: [{ ssrc: '1111' }],
 				rtcp: { cname: 'qwerty' },
 			},
@@ -633,11 +633,11 @@ test('producer.pause() and resume() succeed', async () => {
 
 	// Even if we don't await for pause()/resume() completion, the observer must
 	// fire 'pause' and 'resume' events if state was the opposite.
-	audioProducer.pause();
-	audioProducer.resume();
-	audioProducer.pause();
-	audioProducer.pause();
-	audioProducer.pause();
+	void audioProducer.pause();
+	void audioProducer.resume();
+	void audioProducer.pause();
+	void audioProducer.pause();
+	void audioProducer.pause();
 	await audioProducer.resume();
 
 	expect(onObserverPause).toHaveBeenCalledTimes(3);
@@ -686,7 +686,7 @@ test('producer.enableTraceEvent() succeed', async () => {
 
 	expect(dump2.traceEventTypes).toEqual(expect.arrayContaining([]));
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await audioProducer.enableTraceEvent(['nack', 'FOO', 'fir']);
 
 	const dump3 = await audioProducer.dump();
@@ -707,16 +707,16 @@ test('producer.enableTraceEvent() with wrong arguments rejects with TypeError', 
 		ctx.audioProducerOptions
 	);
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await expect(audioProducer.enableTraceEvent(123)).rejects.toThrow(TypeError);
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await expect(audioProducer.enableTraceEvent('rtp')).rejects.toThrow(
 		TypeError
 	);
 
 	await expect(
-		// @ts-ignore
+		// @ts-expect-error --- Testing purposes.
 		audioProducer.enableTraceEvent(['fir', 123.123])
 	).rejects.toThrow(TypeError);
 }, 2000);
@@ -806,11 +806,8 @@ test('Producer methods reject if closed', async () => {
 	audioProducer.close();
 
 	await expect(audioProducer.dump()).rejects.toThrow(Error);
-
 	await expect(audioProducer.getStats()).rejects.toThrow(Error);
-
 	await expect(audioProducer.pause()).rejects.toThrow(Error);
-
 	await expect(audioProducer.resume()).rejects.toThrow(Error);
 }, 2000);