+++ node/src/test/test-WebRtcTransport.ts
@@ -279,7 +279,7 @@ test('router.createWebRtcTransport() with portRange succeeds', async () => {
 }, 2000);
 
 test('router.createWebRtcTransport() with wrong arguments rejects with TypeError', async () => {
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await expect(ctx.router!.createWebRtcTransport({})).rejects.toThrow(
 		TypeError
 	);
@@ -297,24 +297,24 @@ test('router.createWebRtcTransport() with wrong arguments rejects with TypeError
 	).rejects.toThrow(TypeError);
 
 	await expect(
-		// @ts-ignore
+		// @ts-expect-error --- Testing purposes.
 		ctx.router!.createWebRtcTransport({ listenIps: [123] })
 	).rejects.toThrow(TypeError);
 
 	await expect(
-		// @ts-ignore
+		// @ts-expect-error --- Testing purposes.
 		ctx.router!.createWebRtcTransport({ listenInfos: '127.0.0.1' })
 	).rejects.toThrow(TypeError);
 
 	await expect(
-		// @ts-ignore
+		// @ts-expect-error --- Testing purposes.
 		ctx.router!.createWebRtcTransport({ listenIps: '127.0.0.1' })
 	).rejects.toThrow(TypeError);
 
 	await expect(
 		ctx.router!.createWebRtcTransport({
 			listenIps: ['127.0.0.1'],
-			// @ts-ignore
+			// @ts-expect-error --- Testing purposes.
 			appData: 'NOT-AN-OBJECT',
 		})
 	).rejects.toThrow(TypeError);
@@ -323,7 +323,7 @@ test('router.createWebRtcTransport() with wrong arguments rejects with TypeError
 		ctx.router!.createWebRtcTransport({
 			listenIps: ['127.0.0.1'],
 			enableSctp: true,
-			// @ts-ignore
+			// @ts-expect-error --- Testing purposes.
 			numSctpStreams: 'foo',
 		})
 	).rejects.toThrow(TypeError);
@@ -433,13 +433,13 @@ test('webRtcTransport.connect() with wrong arguments rejects with TypeError', as
 
 	let dtlsRemoteParameters: mediasoup.types.DtlsParameters;
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await expect(webRtcTransport.connect({})).rejects.toThrow(TypeError);
 
 	dtlsRemoteParameters = {
 		fingerprints: [
 			{
-				// @ts-ignore.
+				// @ts-expect-error --- Testing purposes..
 				algorithm: 'sha-256000',
 				value:
 					'82:5A:68:3D:36:C3:0A:DE:AF:E7:32:43:D2:88:83:57:AC:2D:65:E5:80:C4:B6:FB:AF:1A:A0:21:9F:6D:0C:AD',
@@ -460,7 +460,7 @@ test('webRtcTransport.connect() with wrong arguments rejects with TypeError', as
 					'82:5A:68:3D:36:C3:0A:DE:AF:E7:32:43:D2:88:83:57:AC:2D:65:E5:80:C4:B6:FB:AF:1A:A0:21:9F:6D:0C:AD',
 			},
 		],
-		// @ts-ignore
+		// @ts-expect-error --- Testing purposes.
 		role: 'chicken',
 	};
 
@@ -614,7 +614,7 @@ test('transport.enableTraceEvent() succeed', async () => {
 		],
 	});
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await webRtcTransport.enableTraceEvent(['foo', 'probation']);
 	await expect(webRtcTransport.dump()).resolves.toMatchObject({
 		traceEventTypes: ['probation'],
@@ -625,7 +625,7 @@ test('transport.enableTraceEvent() succeed', async () => {
 		traceEventTypes: [],
 	});
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await webRtcTransport.enableTraceEvent(['probation', 'FOO', 'bwe', 'BAR']);
 	await expect(webRtcTransport.dump()).resolves.toMatchObject({
 		traceEventTypes: ['probation', 'bwe'],
@@ -644,18 +644,18 @@ test('transport.enableTraceEvent() with wrong arguments rejects with TypeError',
 		],
 	});
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await expect(webRtcTransport.enableTraceEvent(123)).rejects.toThrow(
 		TypeError
 	);
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await expect(webRtcTransport.enableTraceEvent('probation')).rejects.toThrow(
 		TypeError
 	);
 
 	await expect(
-		// @ts-ignore
+		// @ts-expect-error --- Testing purposes.
 		webRtcTransport.enableTraceEvent(['probation', 123.123])
 	).rejects.toThrow(TypeError);
 }, 2000);
@@ -816,7 +816,7 @@ test('WebRtcTransport methods reject if closed', async () => {
 
 	await expect(webRtcTransport.getStats()).rejects.toThrow(Error);
 
-	// @ts-ignore
+	// @ts-expect-error --- Testing purposes.
 	await expect(webRtcTransport.connect({})).rejects.toThrow(Error);
 
 	await expect(webRtcTransport.setMaxIncomingBitrate(200000)).rejects.toThrow(